+++ core/src/main/java/org/jdbi/v3/core/h2/H2DatabasePlugin.java
@@ -13,8 +13,6 @@
  */
 package org.jdbi.v3.core.h2;
 
-import java.util.UUID;
-
 import org.jdbi.v3.core.Jdbi;
 import org.jdbi.v3.core.array.SqlArrayArgumentStrategy;
 import org.jdbi.v3.core.spi.JdbiPlugin;
@@ -26,18 +24,5 @@ public class H2DatabasePlugin extends JdbiPlugin.Singleton {
     @Override
     public void customizeJdbi(Jdbi db) {
         db.setSqlArrayArgumentStrategy(SqlArrayArgumentStrategy.OBJECT_ARRAY);
-        db.registerArrayType(UUID.class, "uuid")
-            .registerArrayType(short.class, "smallint")
-            .registerArrayType(Short.class, "smallint")
-            .registerArrayType(int.class, "integer")
-            .registerArrayType(Integer.class, "integer")
-            .registerArrayType(long.class, "bigint")
-            .registerArrayType(Long.class, "bigint")
-            .registerArrayType(String.class, "varchar")
-            .registerArrayType(UUID.class, "uuid")
-            .registerArrayType(float.class, "float4")
-            .registerArrayType(Float.class, "float4")
-            .registerArrayType(double.class, "float8")
-            .registerArrayType(Double.class, "float8");
     }
 }